fix(lab-new/02): add more contents to lab report
