+++ src/lab-new/02/assets/showPCB/Makefile
@@ -1,0 +1,10 @@
+obj-m += showPCB.o
+
+PWD := $(CURDIR)
+BUILDDIR := /lib/modules/$(shell uname -r)/build
+
+all:
+	make -C $(BUILDDIR) M=$(PWD) modules 
+
+clean: 
+	make -C $(BUILDDIR) M=$(PWD) clean
